Installing the latest version of the plugin
The previous version has a bug that makes cluster icons to malfunction
for keyboard users. The version 1.5.1 fixes that bug:

https://github.com/Leaflet/Leaflet.markercluster/blob/master/CHANGELOG.md#151-2021-08-28

Since version 1.5.0 releases does not include the file `dist/leaflet.markercluster.js`,
therefore the asset pipeline does not find the file and crashes during assets
precompilation. So I included the missing file for the `1.5.3` version in the vendor
folder by dowloading it from the npm CDN

* https://github.com/Leaflet/Leaflet.markercluster#using-the-plugin
* https://github.com/Leaflet/Leaflet.markercluster/tags